ai/findingnehair/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.6
numpy==1.24.4
pandas==2.2.3
selenium==4.29.0
python-multipart
efficientnet_pytorch==0.7.1
scikit-learn==1.2.2
pgvector==0.4.0
psycopg2-binary    
pillow==9.4.0   
